Refactor: break model into several files
